requirements.txt
Django==5.1.7
SQLAlchemy
django-environ
psycopg2
djangorestframework
drf-yasg
Pillow
djangorestframework-simplejwt